update scripts how they should have been ran
